Int tests for skipping decorator
