+++ plans/integration/add-possibility-to-skip-functions/main.fmf
@@ -1,0 +1,9 @@
+discover+:
+    filter:
+        - 'tag: add-possibility-to-skip-functions'
+        - 'tag: centos8'
+
+provision:
+    how: libvirt
+    origin_vm_name: c2r_centos8_template
+    develop: true
