+++ tests/integration/inhibit-if-kmods-is-not-supported/test_kmods_not_supported.py
@@ -19,22 +19,8 @@ def test_good_convertion(convert2rhel):
     assert c2r.exitstatus == 0
 
 
-@pytest.fixture()
-def insert_custom_kmod(shell):
-    def factory():
-        origin_kmod_loc = Path("/lib/modules/$(uname -r)/kernel/drivers/net/bonding/bonding.ko.xz")
-        new_kmod_dir = origin_kmod_loc.parent / "custom_module_location"
-
-        shell(f"mkdir {new_kmod_dir.as_posix()}")
-        shell(f"mv {origin_kmod_loc.as_posix()} {new_kmod_dir.as_posix()}")
-        shell(f"insmod {(new_kmod_dir / origin_kmod_loc.name).as_posix()}")
-
-    return factory
-
-
 @pytest.mark.bad_tests
 def test_bad_convertion(shell, insert_custom_kmod, convert2rhel):
-    insert_custom_kmod()
     with convert2rhel(
         ("-y --no-rpm-va --serverurl {} --username {} --password {} --pool {} --debug").format(
             env.str("RHSM_SERVER_URL"),
